edit player form renders when button is hit but does not edit player based on imput.

+++ src/components/Players.js
@@ -1,43 +1,46 @@
-import React from 'react';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
-import { useHistory } from 'react-router-dom';
 import { deletePlayer } from '../api/data/teamData';
+import NewPlayerForm from './NewPlayerForm';
 
-export default function Players({ player, setTeam, setEditPlayer }) {
-  const history = useHistory();
+export default function Players({ player, setTeam, user }) {
+  const [showForm, setShowForm] = useState(false);
 
-  const handleClick = () => {
-    deletePlayer(player.firebaseKey).then(setTeam);
-  };
-
-  const setPlayerEdit = (obj) => {
-    setEditPlayer(obj);
-    history.push('/player');
+  const handleClick = (method) => {
+    if (method === 'delete') {
+      deletePlayer(player.firebaseKey).then(setTeam);
+    } else if (method === 'update') {
+      setShowForm(true);
+    }
   };
 
   return (
     <>
       {/* Need to attempt to put the NewPlayerForm here on button click and use a new Promise to get sinlge player then call the form. */}
+      <hr />
       <i>{player.imageUrl}</i>
       <h3>{player.name}</h3>
-      <hr />
       <h4>{player.position}</h4>
       <div>
         <button
-          onClick={setPlayerEdit(player)}
+          onClick={() => handleClick('update')}
           type="button"
           className="btn btn-info"
         >
           Edit Player
         </button>
         <button
-          onClick={() => handleClick(player)}
+          onClick={() => handleClick('delete')}
           type="button"
           className="btn btn-info"
         >
           Delete Player
         </button>
+        <hr />
       </div>
+      {showForm && (
+        <NewPlayerForm player={player} setTeam={setTeam} user={user} />
+      )}
     </>
   );
 }
@@ -52,7 +55,9 @@ Players.propTypes = {
     uid: PropTypes.string,
   }).isRequired,
   setTeam: PropTypes.func.isRequired,
-  setEditPlayer: PropTypes.func,
+  user: PropTypes.shape({
+    name: PropTypes.string,
+    uid: PropTypes.string,
+  }).isRequired,
+  // setPlayer: PropTypes.func.isRequired,
 };
-
-Players.defaultProps = { setEditPlayer: {} };
